backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==2.7.0
watchdog==4.0.0
PyYAML==6.0.1
